+++ resources/views/fields/select.blade.php
@@ -1,0 +1,16 @@
+<div class="form-group{{ $errors->has($oldName) ? ' has-error' : '' }}">
+    @if(isset($title))
+        <label for="{{$id}}">{{$title}}</label>
+    @endif
+    <select @include('dashboard::partials.fields.attributes', ['attributes' => $attributes])>
+        @foreach($options as $key => $option)
+            <option value="{{$key}}"
+                    @if(isset($value) && $key === $value) selected @endif
+            >{{$option}}</option>
+        @endforeach
+    </select>
+    @if(isset($help))
+        <p class="help-block">{{$help}}</p>
+    @endif
+</div>
+@include('dashboard::partials.fields.hr', ['show' => $hr ?? true])