+++ resources/views/layouts/dashboard.blade.php
@@ -48,6 +48,7 @@
                 <i class="icon-logout"></i>
             </button>
 
+            {{--
             <button class="pull-right visible-xs click"
                     data-turbolinks="false"
                     data-toggle="open"
@@ -55,6 +56,8 @@
                     data-target="#quickview">
                 <i class="icon-bell"></i>
             </button>
+            --}}
+
             <button class="pull-right visible-xs click" data-toggle="open" title="Menu" data-target="#aside">
                 <i class="icon-menu"></i>
             </button>