+++ src/Platform/Http/Controllers/Auth/LoginController.php
@@ -32,6 +32,8 @@ class LoginController extends Controller
      */
     public function __construct()
     {
+        $this->redirectTo = config('platform.prefix');
+
         $this->middleware('guest', ['except' => 'logout']);
     }