+++ config/platform.php
@@ -123,7 +123,8 @@ return [
     |
     */
 
-    'single' => [//App\Core\Behaviors\Single\DemoPage::class,
+    'single' => [
+        Orchid\Core\Behaviors\Single\DemoPage::class,
     ],
 
     /*
@@ -135,7 +136,8 @@ return [
     |
     */
 
-    'many' => [//App\Core\Behaviors\Many\DemoPost::class,
+    'many' => [
+        Orchid\Core\Behaviors\Many\DemoPost::class,
     ],
 
     /*