+++ src/Platform/Core/Behaviors/Many/DemoPost.php
@@ -1,27 +1,31 @@
 <?php
 
-namespace App\Core\Behaviors\Single;
+namespace Orchid\Core\Behaviors\Many;
 
-use Orchid\Platform\Behaviors\Single;
-use Orchid\Platform\Fields\Field;
+use Orchid\Platform\Behaviors\Many;
+use Orchid\Platform\Http\Filters\CreatedFilter;
+use Orchid\Platform\Http\Filters\SearchFilter;
+use Orchid\Platform\Http\Filters\StatusFilter;
+use Orchid\Platform\Http\Forms\Posts\BasePostForm;
 use Orchid\Platform\Http\Forms\Posts\UploadPostForm;
+use Orchid\Platform\Fields\Field;
 
-class DemoPage extends Single
+class DemoPost extends Many
 {
     /**
      * @var string
      */
-    public $name = 'Demo page';
+    public $name = 'Demo post';
 
     /**
      * @var string
      */
-    public $description = 'Demonstrative page';
+    public $description = 'Demonstrative post';
 
     /**
      * @var string
      */
-    public $slug = 'demo-page';
+    public $slug = 'demo';
 
     /**
      * Slug url /news/{name}.
@@ -41,6 +45,20 @@ class DemoPage extends Single
             'id'             => 'sometimes|integer|unique:posts',
             'content.*.name' => 'required|string',
             'content.*.body' => 'required|string',
+        ];
+    }
+
+    /**
+     * HTTP data filters
+     *
+     * @return array
+     */
+    public function filters() : array
+    {
+        return [
+            SearchFilter::class,
+            StatusFilter::class,
+            CreatedFilter::class,
         ];
     }
 
@@ -150,7 +168,20 @@ class DemoPage extends Single
     public function modules() : array
     {
         return [
+            BasePostForm::class,
             UploadPostForm::class,
+        ];
+    }
+
+    /**
+     * Grid View for post type.
+     */
+    public function grid() : array
+    {
+        return [
+            'name'       => 'Name',
+            'publish_at' => 'Date of publication',
+            'created_at' => 'Date of creation',
         ];
     }
 }