+++ src/Platform/Providers/FoundationServiceProvider.php
@@ -21,7 +21,6 @@ class FoundationServiceProvider extends ServiceProvider
         $this->registerTranslations();
         $this->registerConfig();
         $this->registerViews();
-        $this->registerCode();
 
         $this->registerProviders();
     }
@@ -72,16 +71,6 @@ class FoundationServiceProvider extends ServiceProvider
         ]), 'dashboard');
     }
 
-    /**
-     * Register types.
-     */
-    protected function registerCode()
-    {
-        $this->publishes([
-            realpath(DASHBOARD_PATH.'/resources/stubs/behaviors/DemoPost.stub') => app_path('Core/Behaviors/Many/DemoPost.php'),
-            realpath(DASHBOARD_PATH.'/resources/stubs/behaviors/DemoPage.stub') => app_path('Core/Behaviors/Single/DemoPage.php'),
-        ]);
-    }
 
     public function registerProviders()
     {